+++ src/main/java/c4/champions/common/affix/affix/AffixScrapper.java
@@ -22,13 +22,11 @@ package c4.champions.common.affix.affix;
 import c4.champions.common.affix.core.AffixBase;
 import c4.champions.common.affix.core.AffixCategory;
 import c4.champions.common.capability.IChampionship;
-import c4.champions.common.config.ConfigHandler;
+import c4.champions.common.ConfigHandler;
 import c4.champions.common.init.ChampionsRegistry;
 import net.minecraft.entity.EntityLiving;
-import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.monster.EntityCreeper;
 import net.minecraft.potion.PotionEffect;
-import net.minecraft.util.DamageSource;
 import net.minecraftforge.event.entity.living.LivingAttackEvent;
 
 public class AffixScrapper extends AffixBase {
@@ -38,11 +36,10 @@ public class AffixScrapper extends AffixBase {
     }
 
     @Override
-    public void onAttack(EntityLiving entity, IChampionship cap, EntityLivingBase target, DamageSource source, float
-            amount, LivingAttackEvent evt) {
+    public void onAttack(EntityLiving entity, IChampionship cap, LivingAttackEvent evt) {
 
         if (!entity.world.isRemote && entity.getRNG().nextFloat() < ConfigHandler.affix.scrapper.chance) {
-            target.addPotionEffect(new PotionEffect(ChampionsRegistry.injured, 200, 0));
+            evt.getEntityLiving().addPotionEffect(new PotionEffect(ChampionsRegistry.injured, 200, 0));
         }
     }