+++ src/main/java/c4/champions/common/capability/CapabilityChampionship.java
@@ -21,7 +21,7 @@ package c4.champions.common.capability;
 
 import c4.champions.Champions;
 import c4.champions.common.affix.EnumAffix;
-import c4.champions.common.config.ConfigHandler;
+import c4.champions.common.ConfigHandler;
 import c4.champions.common.rank.Rank;
 import c4.champions.common.rank.RankManager;
 import c4.champions.common.util.ChampionHelper;