+++ src/main/java/c4/champions/integrations/scalinghealth/ChampionDifficulty.java
@@ -19,7 +19,7 @@
 
 package c4.champions.integrations.scalinghealth;
 
-import c4.champions.common.config.ConfigHandler;
+import c4.champions.common.ConfigHandler;
 
 import java.util.Map;
 import java.util.TreeMap;